Split jest env from utilities for resolver purposes
