+++ packages/sync-server/__tests__/utilities.js
@@ -2,89 +2,15 @@ import config from 'config';
 import supertest from 'supertest';
 import Chance from 'chance';
 import http from 'http';
-import 'jest-expect-message';
-import * as jestExtendedMatchers from 'jest-extended';
 
-import { COMMUNICATION_STATUSES, JWT_TOKEN_TYPES } from 'shared/constants';
+import { COMMUNICATION_STATUSES, JWT_TOKEN_TYPES } from '@tamanu/shared/constants';
 import { createApp } from 'sync-server/app/createApp';
 import { initDatabase, closeDatabase } from 'sync-server/app/database';
 import { getToken } from 'sync-server/app/auth/utils';
 import { DEFAULT_JWT_SECRET } from 'sync-server/app/auth';
 import { initIntegrations } from 'sync-server/app/integrations';
 
-jest.setTimeout(30 * 1000); // more generous than the default 5s but not crazy
-jest.mock('../app/utils/getFreeDiskSpace');
-
 const chance = new Chance();
-
-const formatError = response => {
-  if (!response.body) {
-    return `
-
-Error has no body! (Did you forget to await?)
-`;
-  }
-  return `
-
-Error details:
-${JSON.stringify(response.body.error, null, 2)}
-`;
-};
-
-export function extendExpect(expect) {
-  // Needs to be added explicitly because of the jest-expect-message import
-  expect.extend(jestExtendedMatchers);
-  expect.extend({
-    toBeForbidden(response) {
-      const { statusCode } = response;
-      const pass = statusCode === 403;
-      if (pass) {
-        return {
-          message: () =>
-            `Expected not forbidden (!== 403), got ${statusCode}. ${formatError(response)}`,
-          pass,
-        };
-      }
-      return {
-        message: () => `Expected forbidden (403), got ${statusCode}. ${formatError(response)}`,
-        pass,
-      };
-    },
-    toHaveRequestError(response, expected) {
-      const { statusCode } = response;
-      const match = expected === statusCode;
-      const pass = statusCode >= 400 && statusCode !== 403 && (statusCode < 500 || match);
-      let expectedText = 'Expected error status code';
-      if (expected) {
-        expectedText += ` ${expected}`;
-      }
-      if (pass) {
-        return {
-          message: () => `${expectedText}, got ${statusCode}.`,
-          pass,
-        };
-      }
-      return {
-        message: () => `${expectedText}, got ${statusCode}. ${formatError(response)}`,
-        pass,
-      };
-    },
-    toHaveSucceeded(response) {
-      const { statusCode } = response;
-      const pass = statusCode < 400;
-      if (pass) {
-        return {
-          message: () => `Expected failure status code, got ${statusCode}.`,
-          pass,
-        };
-      }
-      return {
-        message: () => `Expected success status code, got ${statusCode}. ${formatError(response)}`,
-        pass,
-      };
-    },
-  });
-}
 
 class MockApplicationContext {
   closeHooks = [];
